Develop

+++ GieldaL2.INFRASTRUCTURE/DTO/StatisticsDTO.cs
@@ -2,18 +2,18 @@
 {
     public class StatisticsDTO
     {
-        public float BackendTime { get; set; }
+        public int BackendTime { get; set; }
 
-        public float SelectsTime { get; set; }
+        public int SelectsTime { get; set; }
         public int SelectsCount { get; set; }
 
-        public float UpdatesTime { get; set; }
+        public int UpdatesTime { get; set; }
         public int UpdatesCount { get; set; }
 
-        public float InsertsTime { get; set; }
+        public int InsertsTime { get; set; }
         public int InsertsCount { get; set; }
 
-        public float DeletesTime { get; set; }
+        public int DeletesTime { get; set; }
         public int DeletesCount { get; set; }
     }
 }
